+++ src/lib.rs
@@ -170,6 +170,7 @@
 //! `skip` and `skip_inner`. To prevent breaking invariants defined for these
 //! traits, some of them can only be skipped in groups. The following groups are
 //! available:
+//! - [`Clone`]: Uses [`Default`] instead of [`Clone`].
 //! - [`Debug`]
 //! - `EqHashOrd`: Skips [`Eq`], [`Hash`], [`Ord`], [`PartialOrd`] and
 //!   [`PartialEq`].