+++ tests/ui/zeroize/item_skip.rs
@@ -2,8 +2,8 @@ use std::marker::PhantomData;
 
 use derive_where::derive_where;
 
-#[derive_where(Clone; T)]
-#[derive_where(skip_inner(Clone))]
+#[derive_where(Copy; T)]
+#[derive_where(skip_inner(Copy))]
 struct UnsupportedTrait<T>(PhantomData<T>);
 
 fn main() {}