+++ tests/ui/zeroize/skip.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 
 use derive_where::derive_where;
 
-#[derive_where(Clone; T)]
-struct UnsupportedTrait<T>(#[derive_where(skip(Clone))] PhantomData<T>);
+#[derive_where(Skip; T)]
+struct UnsupportedTrait<T>(#[derive_where(skip(Skip))] PhantomData<T>);
 
 fn main() {}