+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Allow skipping fields for `Clone`, calling `Default::default()` instead.
+  **Note:** `Clone` is excluded from blanket skipping and can only be used with
+  selective skipping to avoid this being a breaking change.
+
 ## [1.3.0] - 2025-04-21
 
 ### Added